+++ include/mneme/MnemeUtils.hpp
@@ -3,8 +3,6 @@
 #include <cstdint>
 #include <iomanip>
 #include <iostream>
-#include <llvm/ADT/StringRef.h>
-#include <llvm/Support/raw_ostream.h>
 #include <string>
 
 #if MNEME_ENABLE_DEBUG
@@ -27,16 +25,6 @@ template <typename Ty> Ty extractScalar(char *&Buffer) {
   Ty Value = *reinterpret_cast<const Ty *>(Buffer);
   Buffer += sizeof(Ty);
   return Value;
-}
-
-template <typename Ty>
-void writeScalar(llvm::raw_ostream &OS, const Ty &Value) {
-  OS << llvm::StringRef(reinterpret_cast<const char *>(&Value), sizeof(Value));
-}
-
-inline void writeBytes(llvm::raw_ostream &OS, const void *Data, size_t Size) {
-  if (Size)
-    OS << llvm::StringRef(reinterpret_cast<const char *>(Data), Size);
 }
 
 inline std::string readSizedString(const char *&Buffer) {
